Rework the way comments interact with multiline
Comment printing used to depend on whether the code was multiline. Now
this dependency is turned around: whether something will be rendered on
multiple lines depends on whether there are comments. This makes for a
much more natural way of formatting code.

This also adds support for comments next to function arguments.